+++ docs/version/changelog.md
@@ -5,9 +5,12 @@
 Feat:
 
 * Added stacked bar chart Outdoor Comfort
+* Added filter for the descriptive statistics table in the Data Explorer tab
 
 Fix:
-
+* Fixed issues 154, 156: Update the option list 
+* Fixed issues 157: Align the legend bar of "Thermal Stress" with the text
+* Fixed issues 122
 * Fixed issues 144, 148, and 141
 * Correctly aligned SI and IP buttons
 * Fixed typos in variables